Cleanup events names to avoid any conflict between tray-dashboard-renderer

+++ src/global/json.ts
@@ -1,6 +1,7 @@
 const { existsSync, readFileSync, writeFileSync } = require('fs');
 
 export const readJson = (filename: string, isArray: boolean = false): { exist: boolean; valid: boolean; data: any } => {
+	if (typeof filename !== 'string' || !filename.endsWith('.json')) return { exist: false, valid: false, data: null };
 	try {
 		if (!existsSync(filename)) return { exist: false, valid: false, data: null };
 		const data = JSON.parse(readFileSync(filename, 'utf-8'));
@@ -13,6 +14,7 @@ export const readJson = (filename: string, isArray: boolean = false): { exist: b
 };
 
 export const writeJSON = (filename: string, data: any): { exist: boolean; valid: boolean; data: any } => {
+	if (typeof filename !== 'string' || !filename.endsWith('.json')) return { exist: false, valid: false, data: null };
 	if (!data || typeof data !== 'object') return { exist: false, valid: false, data: null };
 	try {
 		writeFileSync(filename, JSON.stringify(data, null, 2) + '\n');
